Stable implementation of genetic algorithm

+++ home.py
@@ -644,12 +644,12 @@ def run_simulation_click(boid_profile, ga_profile, show_graphics):
 
     boid_simulation.run(
         num_boids           = int(boid_config[boid_profile_name]['num_boids']),
-        max_speed           = int(boid_config[boid_profile_name]['max_speed']), 
-        neighbor_radius     = int(boid_config[boid_profile_name]['neighbor_radius']),
+        max_speed           = float(boid_config[boid_profile_name]['max_speed']), 
+        neighbor_radius     = float(boid_config[boid_profile_name]['neighbor_radius']),
         alignment_weight    = float(boid_config[boid_profile_name]['alignment_weight']),
         cohesion_weight     = float(boid_config[boid_profile_name]['cohesion_weight']),
         separation_weight   = float(boid_config[boid_profile_name]['separation_weight']),
-        avoid_radius        = int(boid_config[boid_profile_name]['avoid_radius']),
+        avoid_radius        = float(boid_config[boid_profile_name]['avoid_radius']),
         max_avoid_force     = float(boid_config[boid_profile_name]['max_avoid_force']),
         show_graphics       = show_graphics.get()
     )
